16k: Statically align PT_LOAD Program headers to max-page-size boundary
When the page size is 4k and shared libraries and binaries are
16k/64k elf aligned with the flag -Wl,-z,max-page-size=[16384|65536],
the dynamic linker (loader) does not unmap the hole between segments an
a extra vm_area_struct is created.

This happens because the loader allocates a memory area big enough to
map the shared library and then maps and mprotect every segment at
page size boundaries instead of a p_align boundary.

This program implements a PoC of that extends the p_filesz and p_memsz
to a p_align boundary statically. This can be achieved by
adding a new elf64 section to each PT_LOAD segment to fill the hole.

Bug: 311934502
Change-Id: I67dc3f35994e87c79d29002577310c999ed00778